requirements/dev.txt
-r base.txt

django_settings @ file:../django-settings-class